pyproject.toml
[project]
name = "studybuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.0",
    "openai-agents>=0.1.0",
    "pandas>=2.3.0",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "pypdf2>=3.0.0",
    "rich>=14.0.0",
    "fpdf>=1.7.2",
]
